Type i32 -> isize

+++ libgame/src/battle.rs
@@ -5,7 +5,7 @@ use serde::{Deserialize, Serialize};
 
 use indexmap::IndexMap;
 
-type Player = i32;
+type Player = isize;
 
 pub trait Read {
     fn read_line(&mut self, buffer: &mut String) -> io::Result<usize>;
@@ -23,7 +23,7 @@ struct PlayData {
     state: &'static str,
     player: Option<Player>,
     data: Option<&'static str>,
-    points: Option<Vec<i32>>
+    points: Option<Vec<isize>>
 }
 
 #[derive(PartialEq, Debug, Serialize, Deserialize)]
@@ -34,15 +34,15 @@ pub struct PlayerMove {
 
 #[derive(PartialEq, Debug)]
 pub struct Battle<R: Read, W: Write> {
-    pub num_players: i32,
-    points: IndexMap<Player, i32>,
+    pub num_players: isize,
+    points: IndexMap<Player, isize>,
 
     stdin: R,
     stdout: W
 }
 
 impl<R: Read, T: Write> Battle<R, T> {
-    pub fn new(num_players: i32) -> Battle<Stdin, Stdout> {
+    pub fn new(num_players: isize) -> Battle<Stdin, Stdout> {
         let mut points = IndexMap::new();
         for player in 0..num_players {
             points.insert(player+1, 0);
@@ -83,11 +83,11 @@ impl<R: Read, T: Write> Battle<R, T> {
         }
     }
 
-    pub fn set_points(&mut self, player: Player, count: i32) {
+    pub fn set_points(&mut self, player: Player, count: isize) {
         self.points.insert(player, count);
     }
 
-    pub fn add_points(&mut self, player: Player, count: i32) {
+    pub fn add_points(&mut self, player: Player, count: isize) {
         let points = self.points.get(&player).unwrap_or(&0);
         self.points.insert(player, points + count);
     }
@@ -96,7 +96,7 @@ impl<R: Read, T: Write> Battle<R, T> {
         self.points.sort_keys();
         let play_data = PlayData {
             state: "end",
-            points: Some(self.points.values().cloned().collect::<Vec<i32>>()),
+            points: Some(self.points.values().cloned().collect::<Vec<isize>>()),
             player: None,
             data
         };
